requirements.txt
playwright==1.49.1
python-dotenv==1.0.1
rich==13.9.4
fastapi==0.115.0
uvicorn==0.31.1
websockets==13.1
